requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.6.0
google-api-python-client>=2.70.0
Pillow>=9.3.0
torch>=2.0.0
torchvision>=0.15.0
pytesseract>=0.3.10
numpy>=1.21.0
tqdm>=4.65.0
python-dotenv>=0.19.0
pathlib>=1.0.1
ratelimit>=2.2.1
requests>=2.31.0 